Move FilterWidget into filters directory
