+++ resources/frontend_client/app/query_builder/GuiQueryEditor.react.js
@@ -3,7 +3,7 @@
 import AggregationWidget from './AggregationWidget.react';
 import DataSelector from './DataSelector.react';
 import FieldWidget from './FieldWidget.react';
-import FilterWidget from './FilterWidget.react';
+import FilterWidget from './filters/FilterWidget.react';
 import FilterPopover from './filters/FilterPopover.react';
 import Icon from "metabase/components/Icon.react";
 import IconBorder from 'metabase/components/IconBorder.react';
